move upload functions to keygen.cpp
