+++ src/cli.cpp
@@ -8,16 +8,12 @@
 #include <iostream>
 #include <filesystem>
 #include <fstream>
-#include <curl/curl.h>
 
 #ifdef ENABLE_TESTS
 #include <gtest/gtest.h>
 #endif
 
 namespace fs = std::filesystem;
-
-static const char* FLASK_PORT = "5000";
-static long curlConnectionTimeout = 10L;
 
 void printUsage(std::string mode) {
     if (mode == "generate") {
@@ -36,46 +32,6 @@ void printUsage(std::string mode) {
 #endif
     } else {
         std::cout << GeneralUsage;
-    }
-}
-
-static size_t curlWriteCallback(char* data, size_t size, size_t nmemb, std::string* response) {
-    size_t totalSize = size * nmemb;
-    response->append(data, totalSize);
-    return totalSize;
-}
-
-static void uploadFileToCodespace(const std::string& filename, const std::string& codespaceName) {
-    std::string url = "https://" + codespaceName + "-" + FLASK_PORT +  ".preview.app.github.dev/upload";
-
-    CURL* curl = curl_easy_init();
-    if (curl) {
-        curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
-        curl_easy_setopt(curl, CURLOPT_TIMEOUT, curlConnectionTimeout);
-
-        // Set the multipart/form-data request
-        curl_mime* mime = curl_mime_init(curl);
-        curl_mimepart* part = curl_mime_addpart(mime);
-        curl_mime_name(part, "file");
-        curl_mime_filedata(part, filename.c_str());
-
-        curl_easy_setopt(curl, CURLOPT_MIMEPOST, mime);
-
-        std::string serverResponse;
-        curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, curlWriteCallback);
-        curl_easy_setopt(curl, CURLOPT_WRITEDATA, &serverResponse);
-
-        CURLcode res = curl_easy_perform(curl);
-        if (res == CURLE_OK) {
-            std::cout << "File uploaded successfully to the remote codespace at " << serverResponse << std::endl;
-        } else {
-            throw std::invalid_argument("Error: " + std::string(curl_easy_strerror(res)));
-        }
-
-        curl_mime_free(mime);
-        curl_easy_cleanup(curl);
-    } else {
-        throw std::invalid_argument("Failed to initialize libcurl");
     }
 }
 
@@ -206,6 +162,9 @@ int main(int argc, char* argv[]) {
         printUsage(mode);
         std::cout << "\nERROR: " << ex.what() << std::endl << std::endl;
         return 1;
+    }
+    catch (const std::exception& ex) {
+        std::cout << "\nERROR: " << ex.what() << std::endl << std::endl;
     }
     catch (QryptSecurity::QryptSecurityException& ex) {
         std::cout << "\nSDK ERROR: " << ex.what() << std::endl << std::endl;
